requirements.txt
gym==0.19.0
JSBSim==1.1.11
matplotlib==3.5.1
numpy==1.22.2
pyttsx3==2.90
PyYAML==6.0
setuptools==57.0.0
stable_baselines3==1.4.0
torch==1.10.2
tensorboard==2.8.0
